+++ view/menuutama.php
@@ -35,10 +35,10 @@
                     <li><a href="?page=pendaftaran&form=spp">Data Preparasi</a></li>
                 </ul>
             </li>
-            <li> <a href="javascript:void(0);" class="menu-toggle"><i class="zmdi zmdi-assignment"></i><span>Transaksi</span></a>
+            <li> <a href="javascript:void(0);" class="menu-toggle"><i class="zmdi zmdi-assignment"></i><span>Master Data</span></a>
                 <ul class="ml-menu">
                     <li><a href="?page=pendaftaran&form=spp">SPP</a></li>
-                    <li><a href="?page=pendaftaran&form=transaksi">Logbook</a></li>
+                    <li><a href="?page=pendaftaran&form=logbook">Logbook</a></li>
                 </ul>
             </li>
             <li> <a href="javascript:void(0);" class="menu-toggle"><i class="zmdi zmdi-account"></i><span>Pegawai</span></a>
